Fixed task size input.
The input type=number was problematic in a number of ways.

I created a simple Select component and changed size to be a select of
1, 3, 5, 8 and 13.  In addition to saner input, this is saner
semantically. (Tasks bigger than 13 points need to be broken apart. :) .)

I also combined the size, assigned and blocked input into a single row
(where there's room) to make the task dialog more compact.